requirements.txt
pyspark: pip install pyspark
pyyaml: pip install pyyaml